[ELY-1736] Fork 'org/wildfly/security/password/util/PasswordUtil' where appropriate and move the class to 'wildfly-elytron-password-impl'
