+++ password/impl/src/main/java/org/wildfly/security/password/impl/SaltedSimpleDigestPasswordImpl.java
@@ -31,7 +31,6 @@ import java.security.spec.InvalidKeySpecException;
 import java.security.spec.KeySpec;
 import java.util.Arrays;
 
-import org.wildfly.security.password.util.PasswordUtil;
 import org.wildfly.security.password.interfaces.SaltedSimpleDigestPassword;
 import org.wildfly.security.password.spec.ClearPasswordSpec;
 import org.wildfly.security.password.spec.SaltedHashPasswordSpec;
