+++ password/impl/src/main/java/org/wildfly/security/password/impl/ScramDigestPasswordImpl.java
@@ -39,7 +39,6 @@ import javax.crypto.spec.SecretKeySpec;
 import org.wildfly.security.password.Password;
 import org.wildfly.security.password.spec.IteratedPasswordAlgorithmSpec;
 import org.wildfly.security.password.spec.SaltedPasswordAlgorithmSpec;
-import org.wildfly.security.password.util.PasswordUtil;
 import org.wildfly.security.password.interfaces.ScramDigestPassword;
 import org.wildfly.security.password.spec.ClearPasswordSpec;
 import org.wildfly.security.password.spec.IteratedSaltedPasswordAlgorithmSpec;
